// src/types/config.ts
import { PaletteMode } from '@mui/material';

export type ConfigProps = {
  layout: string;
  drawerType: string;
  fontFamily: string;
  borderRadius: number;
  outlinedFilled: boolean;
  navType: PaletteMode;
  presetColor: string;
  locale: string;
  rtlLayout: boolean;
  container: boolean;
  dashoardPath: string,
  hMaxItem: number,
  selectedID: string,
  drawerOpen: boolean,
};

export type CustomizationProps = {
  layout: string;
  drawerType: string;
  fontFamily: string;
  borderRadius: number;
  outlinedFilled: boolean;
  navType: PaletteMode;
  presetColor: string;
  locale: string;
  rtlLayout: boolean;
  container: boolean;
  dashoardPath: string,
  hMaxItem: number,
  drawerOpen: boolean,
  selectedID: string,
  onChangeLayout: (layout: string) => void;
  onChangeDrawer: ({ drawerType, drawerOpen }: { drawerType: string, drawerOpen: boolean }) => void;
  onChangeMenuType: (navType: PaletteMode) => void;
  onChangePresetColor: (presetColor: string) => void;
  onChangeLocale: (locale: string) => void;
  onChangeRTL: (rtlLayout: boolean) => void;
  onChangeContainer: () => void;
  onChangeFontFamily: (fontFamily: string) => void;
  onChangeBorderRadius: (event: Event, newValue: number | number[]) => void;
  onChangeOutlinedField: (outlinedFilled: boolean) => void;
  onChangeSelectedID: (selectedID: number) => void;
};